pyproject.toml
[project]
name = "fashion-trend-intelligence"
version = "0.1.0"
description = "segmentation de photos pour identifier vetements"
authors = [
    {name = "Damien GUESDON", email = "kisai.dg.slu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
requests = "^2.32"
python-dotenv = "^1.0"
huggingface-hub = "^0.27"
transformers = "^4.30"
Pillow = "^11.0"
numpy = "^2.0.0"
matplotlib = "^3.8"
pycocotools = "^2.0"
torch = "2.2.2"
pandas = "^2.3.1"
opencv-python = "^4.12.0.88"
jupyterlab = "^4.4.5"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.4.4"
ipykernel  = "^6.29.5"
pytest     = "^8.4.1"
flake8     = "^7.3.0"
black      = "^25.1.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
